backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
pgvector
langchain-google-genai
langchain-community
pandas
python-dotenv
black
redis
langchain-redis
google-api-python-client
google-auth-httplib2
google-auth-oauthlib